Add TestLogger, parse prices as floats, add AssertJ
Introduce a TestLogger utility and replace direct System.out prints with logger calls. Change ItemForSale.getPrice() to return Float and robustly parse localized numeric price strings (with fallbacks). Update CraigListSalePanel sorting helpers to work with List<Float>. Enhance CraigListSalePanelSortButton to log sorting actions, wait briefly after selection, and capture a screenshot to ensure sorting completes. Update tests to use AssertJ assertions and to handle Float prices. Add assertj-core dependency and configure maven-surefire-report-plugin in the POM.

+++ CraiglistAutomation/src/main/java/teckro/testlibraries/businesslogic/CraigListSalePanel.java
@@ -37,22 +37,22 @@ public class CraigListSalePanel {
             .toList();
     }
 
-    public static boolean isAscending(List<Integer> values) {
+    public static boolean isAscending(List<Float> values) {
         if (values.isEmpty()) return true;
-        int prev = values.get(0);
+        Float prev = values.get(0);
         for (int i = 1; i < values.size(); i++) {
-            int current = values.get(i);
+            Float current = values.get(i);
             if (current < prev) return false;
             prev = current;
         }
         return true;
     }
 
-    public static boolean isDescending(List<Integer> values) {
+    public static boolean isDescending(List<Float> values) {
         if (values.isEmpty()) return true;
-        int prev = values.get(0);
+        Float prev = values.get(0);
         for (int i = 1; i < values.size(); i++) {
-            int current = values.get(i);
+            Float current = values.get(i);
             if (current > prev) return false;
             prev = current;
         }
